feat(computer-id): allow "per-device" path.
The plugin have some problem using sync, as the path will be exactly the same even on different computer.
I added the "machine-node-id" dep to generate an id for each device.
Each device will have always different id, so the path will be saved in a "per device" and loaded by the id.
When the path is updated, it is also saved in the devices configuration.